pyproject.toml
[project]
name = "myagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]

[project.scripts]
myagent = "myagent:main"
start-myagent = "myagent.clsAgent:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
